requirements.txt
joblib==1.2.0
matplotlib>3.7.0
numpy>=1.20.0
pandas>=2.2.0
pytest>=7.4.0
scikit_learn==1.2.2
sklearn_genetic_opt>=0.10.1
dill>=0.3.8
Flask>=3.0.2
